chore: Move build editor script to scripts/ folder
