+++ deno.json
@@ -14,7 +14,7 @@
 		"lint": "deno run --allow-env --allow-read --allow-write --allow-run=git scripts/lint.js",
 		"check": "deno run --unstable --allow-env --allow-run --allow-read --allow-write --allow-net scripts/check.js",
 		"test": "deno run --unstable --allow-run --allow-read --allow-write --allow-env --allow-net scripts/test.js",
-		"build-engine": "deno run --unstable --allow-read --allow-write --allow-net --allow-env scripts/build.js",
-		"build-editor": "deno run --unstable -A editor/scripts/build.js"
+		"build-engine": "deno run --unstable --allow-read --allow-write --allow-net --allow-env scripts/buildEngine.js",
+		"build-editor": "deno run --unstable -A scripts/buildEditor.js"
 	}
 }
