+++ scripts/buildEngine.js
